memcache_proto::binary

Enum BinaryCommand

pub enum BinaryCommand<'a> {

    Get {
        key: &'a [u8],
        opaque: u32,
    },
    GetQ {
        key: &'a [u8],
        opaque: u32,
    },
    GetK {
        key: &'a [u8],
        opaque: u32,
    },
    GetKQ {
        key: &'a [u8],
        opaque: u32,
    },
    Set {
        key: &'a [u8],
        value: &'a [u8],
        flags: u32,
        expiration: u32,
        cas: u64,
        opaque: u32,
    },
    SetQ {
        key: &'a [u8],
        value: &'a [u8],
        flags: u32,
        expiration: u32,
        cas: u64,
        opaque: u32,
    },
    Add {
        key: &'a [u8],
        value: &'a [u8],
        flags: u32,
        expiration: u32,
        opaque: u32,
    },
    Replace {
        key: &'a [u8],
        value: &'a [u8],
        flags: u32,
        expiration: u32,
        cas: u64,
        opaque: u32,
    },
    Delete {
        key: &'a [u8],
        cas: u64,
        opaque: u32,
    },
    DeleteQ {
        key: &'a [u8],
        cas: u64,
        opaque: u32,
    },
    Increment {
        key: &'a [u8],
        delta: u64,
        initial: u64,
        expiration: u32,
        cas: u64,
        opaque: u32,
    },
    Decrement {
        key: &'a [u8],
        delta: u64,
        initial: u64,
        expiration: u32,
        cas: u64,
        opaque: u32,
    },
    Append {
        key: &'a [u8],
        value: &'a [u8],
        cas: u64,
        opaque: u32,
    },
    Prepend {
        key: &'a [u8],
        value: &'a [u8],
        cas: u64,
        opaque: u32,
    },
    Touch {
        key: &'a [u8],
        expiration: u32,
        opaque: u32,
    },
    Gat {
        key: &'a [u8],
        expiration: u32,
        opaque: u32,
    },
    Flush {
        expiration: u32,
        opaque: u32,
    },
    Noop {
        opaque: u32,
    },
    Version {
        opaque: u32,
    },
    Quit {
        opaque: u32,
    },
    Stat {
        key: Option<&'a [u8]>,
        opaque: u32,
    },

}

A parsed Memcache binary protocol command.

Variants

Get

GET key

Fields

key: &'a [u8]

opaque: u32

GetQ

GET key (quiet - no response on miss)

Fields

key: &'a [u8]

opaque: u32

GetK

GETK key (returns key in response)

Fields

key: &'a [u8]

opaque: u32

GetKQ

GETKQ key (quiet, returns key in response)

Fields

key: &'a [u8]

opaque: u32

Set

SET key value

Fields

key: &'a [u8]

value: &'a [u8]

flags: u32

expiration: u32

cas: u64

opaque: u32

SetQ

SET key value (quiet - no response on success)

Fields

key: &'a [u8]

value: &'a [u8]

flags: u32

expiration: u32

cas: u64

opaque: u32

Add

ADD key value (only if key doesn’t exist)

Fields

key: &'a [u8]

value: &'a [u8]

flags: u32

expiration: u32

opaque: u32

Replace

REPLACE key value (only if key exists)

Fields

key: &'a [u8]

value: &'a [u8]

flags: u32

expiration: u32

cas: u64

opaque: u32

Delete

DELETE key

Fields

key: &'a [u8]

cas: u64

opaque: u32

DeleteQ

DELETE key (quiet)

Fields

key: &'a [u8]

cas: u64

opaque: u32

Increment

INCREMENT key delta initial expiration

Fields

key: &'a [u8]

delta: u64

initial: u64

expiration: u32

cas: u64

opaque: u32

Decrement

DECREMENT key delta initial expiration

Fields

key: &'a [u8]

delta: u64

initial: u64

expiration: u32

cas: u64

opaque: u32

Append

APPEND key value

Fields

key: &'a [u8]

value: &'a [u8]

cas: u64

opaque: u32

Prepend

PREPEND key value

Fields

key: &'a [u8]

value: &'a [u8]

cas: u64

opaque: u32

Touch

TOUCH key expiration

Fields

key: &'a [u8]

expiration: u32

opaque: u32

Gat

GAT (Get and Touch) key expiration

Fields

key: &'a [u8]

expiration: u32

opaque: u32

Flush

FLUSH [expiration]

Fields

expiration: u32

opaque: u32

Noop

NOOP (used for pipelining)

Fields

opaque: u32

Version

VERSION

Fields

opaque: u32

Quit

QUIT

Fields

opaque: u32

Stat

STAT [key]

Fields

key: Option<&'a [u8]>

opaque: u32

Implementations

impl<'a> BinaryCommand<'a>

pub fn parse(data: &'a [u8]) -> Result<(Self, usize), ParseError>

Parse a binary command from a byte buffer.

Returns the parsed command and the number of bytes consumed.

pub fn is_quiet(&self) -> bool

Returns true if this is a quiet command (no response expected on success/miss).

pub fn opaque(&self) -> u32

Returns the opaque value for this command.