memcache_proto

Enum Request

pub enum Request<'a> {

    Get {
        key: &'a [u8],
    },
    Gets {
        keys: &'a [&'a [u8]],
    },
    Set {
        key: &'a [u8],
        value: &'a [u8],
        flags: u32,
        exptime: u32,
    },
    Add {
        key: &'a [u8],
        value: &'a [u8],
        flags: u32,
        exptime: u32,
    },
    Replace {
        key: &'a [u8],
        value: &'a [u8],
        flags: u32,
        exptime: u32,
    },
    Incr {
        key: &'a [u8],
        delta: u64,
    },
    Decr {
        key: &'a [u8],
        delta: u64,
    },
    Append {
        key: &'a [u8],
        value: &'a [u8],
    },
    Prepend {
        key: &'a [u8],
        value: &'a [u8],
    },
    Cas {
        key: &'a [u8],
        value: &'a [u8],
        flags: u32,
        exptime: u32,
        cas_unique: u64,
    },
    Delete {
        key: &'a [u8],
    },
    FlushAll,
    Version,
    Quit,

}

A request builder for encoding Memcache commands.

Variants

Get

GET command: get <key>\r\n

Fields

key: &'a [u8]

Gets

Multi-GET command: get <key1> <key2> ...\r\n

Fields

keys: &'a [&'a [u8]]

Set

SET command: set <key> <flags> <exptime> <bytes>\r\n<data>\r\n

Fields

key: &'a [u8]

value: &'a [u8]

flags: u32

exptime: u32

Add

ADD command: add <key> <flags> <exptime> <bytes>\r\n<data>\r\n

Stores the item only if the key does not already exist.

Fields

key: &'a [u8]

value: &'a [u8]

flags: u32

exptime: u32

Replace

REPLACE command: replace <key> <flags> <exptime> <bytes>\r\n<data>\r\n

Stores the item only if the key already exists.

Fields

key: &'a [u8]

value: &'a [u8]

flags: u32

exptime: u32

Incr

INCR command: incr <key> <delta>\r\n

Fields

key: &'a [u8]

delta: u64

Decr

DECR command: decr <key> <delta>\r\n

Fields

key: &'a [u8]

delta: u64

Append

APPEND command: append <key> 0 0 <bytes>\r\n<data>\r\n

Appends data to an existing itemโs value.

Fields

key: &'a [u8]

value: &'a [u8]

Prepend

PREPEND command: prepend <key> 0 0 <bytes>\r\n<data>\r\n

Prepends data to an existing itemโs value.

Fields

key: &'a [u8]

value: &'a [u8]

Cas

CAS command: cas <key> <flags> <exptime> <bytes> <cas_unique>\r\n<data>\r\n

Compare-and-swap: stores only if the CAS token matches.

Fields

key: &'a [u8]

value: &'a [u8]

flags: u32

exptime: u32

cas_unique: u64

Delete

DELETE command: delete <key>\r\n

Fields

key: &'a [u8]

FlushAll

FLUSH_ALL command: flush_all\r\n

Version

VERSION command: version\r\n

Quit

QUIT command: quit\r\n

Implementations

impl<'a> Request<'a>

pub fn get(key: &'a [u8]) -> Self

Create a GET request.

pub fn gets(keys: &'a [&'a [u8]]) -> Self

Create a multi-GET request.

pub fn set(key: &'a [u8], value: &'a [u8]) -> SetRequest<'a>

Create a SET request.

pub fn add(key: &'a [u8], value: &'a [u8]) -> AddRequest<'a>

Create an ADD request (store only if key does not exist).

pub fn replace(key: &'a [u8], value: &'a [u8]) -> ReplaceRequest<'a>

Create a REPLACE request (store only if key already exists).

pub fn incr(key: &'a [u8], delta: u64) -> Self

Create an INCR request.

pub fn decr(key: &'a [u8], delta: u64) -> Self

Create a DECR request.

pub fn append(key: &'a [u8], value: &'a [u8]) -> Self

Create an APPEND request.

pub fn prepend(key: &'a [u8], value: &'a [u8]) -> Self

Create a PREPEND request.

pub fn cas(key: &'a [u8], value: &'a [u8], cas_unique: u64) -> Self

Create a CAS (compare-and-swap) request.

pub fn delete(key: &'a [u8]) -> Self

Create a DELETE request.

pub fn flush_all() -> Self

Create a FLUSH_ALL request.

pub fn version() -> Self

Create a VERSION request.

pub fn quit() -> Self

Create a QUIT request.

pub fn encode(&self, buf: &mut [u8]) -> usize

Encode this request into a buffer.

Returns the number of bytes written.